process images

+++ data/resize.py
@@ -8,6 +8,8 @@ import SimpleITK as sitk
 import shutil
 from concurrent.futures import ProcessPoolExecutor
 
+from count_files import count_files_with_suffix
+
 def process(mask_path):
     original_mask_path = mask_path
     image_path = None
@@ -230,14 +232,14 @@ def process(mask_path):
         os.remove(image_path)
         print(f"Removed image file {image_path} from _fixed folder")
     else:
-        print(f"Removed image file {image_path} does not exist.")
+        print(f"image file {image_path} does not exist.")
 
     # remove the mask from the _mask folder if successfully resized.
     if os.path.isfile(original_mask_path):
         os.remove(original_mask_path)
         print(f"Removed image mask file {original_mask_path} from _mask_path folder")
     else:
-        print(f"Removed image mask file {original_mask_path} does not exist.")
+        print(f"image mask file {original_mask_path} does not exist.")
 
     # remove the mask from the merged_mask folder if successfully resized.
     merged_mask = f"/cluster/projects/mcintoshgroup/publicData/CT-RATE-Processed/benchmark/CTRATE_Volumes_raw_h5_fp16_noflip_merged_{phase}_masks/{relative_path}"
@@ -245,7 +247,7 @@ def process(mask_path):
         os.remove(merged_mask)
         print(f"Removed image merged mask file {merged_mask} from _mask_path folder")
     else:
-        print(f"Removed image merged mask file {merged_mask} does not exist.")
+        print(f"image merged mask file {merged_mask} does not exist.")
     
 
 if "__main__" == __name__:
@@ -273,11 +275,23 @@ if "__main__" == __name__:
         for _ in tqdm(executor.map(process, mask_paths), total=len(mask_paths)):
             pass
     
-    # remove the unnecessary directories
-    if os.path.isdir(image_root):
+    ## remove empty directories
+
+    # remove the _fixed image directory
+    if os.path.isdir(image_root) and count_files_with_suffix(image_root, '.nii.gz') == 0:
         shutil.rmtree(image_root)
         print(f"{image_root} removed.")
 
-    if os.path.isdir(mask_root):
+    # remove the _mask image directory
+    if os.path.isdir(mask_root) and count_files_with_suffix(mask_root, '.nii.gz') == 0:
         shutil.rmtree(mask_root)
         print(f"{mask_root} removed.")
+
+    # remove the merged_mask image directory
+    merged_mask_root = f"/cluster/projects/mcintoshgroup/publicData/CT-RATE-Processed/benchmark/CTRATE_Volumes_raw_h5_fp16_noflip_merged_{split}_masks/"
+    if os.path.isdir(merged_mask_root) and count_files_with_suffix(merged_mask_root, '.nii.gz') == 0:
+        shutil.rmtree(merged_mask_root)
+        print(f"{merged_mask_root} removed.")
+    
+    print('finished resize.py script')
+
